First commit
Port from local prototyping repository at cc23e64.

Drop async SYSTICK implementation, since that's not specific to the
chip. There could be a future async-cortex-m crate that provides an
implementation, or something external to this crate. Stick to the
peripherals.

Stub an imxrt106x crate, in the spirit of the split HAL. Next commit
will add a 1010x variant.

Doctests do not yet work. The goal of the split HAL is to make
documentation easier. This will explore how that works in practice.