Add examples for Teensy 4
Tested all but the SPI and UART examples. All look OK, except for the
pipe example. The pipe example's final phase will fail in a release
build, but not in a debug build. Cannot reproduce in my local
prototyping repository. That implementation needs some re-work and
re-evaluation, so I wouldn't be surprised if there's a bug in there.