Add compile-time hints to guid chip integration
I reviewed the changes that I needed to support imxrt101x from the
imxrt106x foundation. When I found an assumption that someone else
should check, I put a compile_error! behind a feature flag. This
might help guide users on the things to check when adding support
for a new chip.