Fix IOMUXC dependency patching
The imxrt-iomuxc 0.1.2 release caused a build conflict in the teensy4
examples. The examples use the teensy4-pins crate, which was bringing
in the 0.1.2 release. But, the async HAL was still using 0.1.1 from
a git repository. This resulted in a "these are not the same types!"
error. I'm guessing it's because they come from completely separate
"registries," but I'm not entirely sure right now.

I cherry-picked the imxrt101x support onto an iomuxc v0.1 branch
fork, so we can continue prototyping and using the async HAL from
master.